blob_exists

+++ modules/src-ctl/src/s3_blob_storage.rs
@@ -28,6 +28,28 @@ impl S3BlobStorage {
             tokio_runtime,
         })
     }
+
+    fn blob_exists(&self, hash: &str) -> Result<bool, String> {
+        let path = self.root.join(hash);
+        let key = path.to_str().unwrap();
+        //we fetch the acl to know if the object exists
+        let req_acl = self
+            .client
+            .get_object_acl()
+            .bucket(&self.bucket_name)
+            .key(key);
+        match self.tokio_runtime.block_on(req_acl.send()) {
+            Ok(_acl) => Ok(true),
+            Err(s3::SdkError::ServiceError { err, raw }) => match err.kind {
+                s3::error::GetObjectAclErrorKind::NoSuchKey(_) => Ok(false),
+                _ => {
+                    let _dummy = raw;
+                    Err(format!("error fetching acl: {}", err))
+                }
+            },
+            Err(e) => Err(format!("error fetching acl: {:?}", e)),
+        }
+    }
 }
 
 impl BlobStorage for S3BlobStorage {
@@ -43,31 +65,8 @@ impl BlobStorage for S3BlobStorage {
     fn write_blob(&self, hash: &str, contents: &[u8]) -> Result<(), String> {
         let path = self.root.join(hash);
         let key = path.to_str().unwrap();
-
-        //todo: move to blob_exists()
-        //we fetch the acl to know if the object exists
-        let req_acl = self
-            .client
-            .get_object_acl()
-            .bucket(&self.bucket_name)
-            .key(key);
-        match self.tokio_runtime.block_on(req_acl.send()) {
-            Ok(_acl) => {
-                //object already uploaded
-                return Ok(());
-            }
-            Err(s3::SdkError::ServiceError { err, raw }) => match err.kind {
-                s3::error::GetObjectAclErrorKind::NoSuchKey(_) => {
-                    //object is not present, carry on
-                }
-                _ => {
-                    let _dummy = raw;
-                    return Err(format!("error fetching acl: {}", err));
-                }
-            },
-            Err(e) => {
-                return Err(format!("error fetching acl: {:?}", e));
-            }
+        if self.blob_exists(hash)? {
+            return Ok(());
         }
 
         let req = self.client.put_object().bucket(&self.bucket_name).key(key);
